+++ components/login.jsx
@@ -10,35 +10,42 @@ import {
   Pressable,
 } from "react-native";
 // import courses from "../course.json";
+import CourseSelect from "./course";
+import { PlayerSelect, PlayerNames } from "./players";
 
 export default function Login(props) {
   const { navigation, route } = props;
   const [buttonTree, setButtonTree] = useState([true, false, false, false]);
   const [playerCount, setPlayerCount] = useState();
   const [players, setPlayers] = useState([]);
-  const [pickedHole, setPickedHole] = useState();
+  const [pickedCourse, setPickedCourse] = useState();
   const [courses, setCourses] = useState([]);
   const [courseLoading, setCourseLoading] = useState(true);
   const [holeDetails, setHoleDetails] = useState([]);
   const [holeLoading, setHoleLoading] = useState(true);
 
+  function resetButtons() {
+    setButtonTree([true, false, false, false]);
+    setPickedCourse();
+    setCourseLoading(true);
+  }
+
   async function courseList() {
-    const courseFetch = await fetch('https://www.overpass-api.de/api/interpreter', {
+    const coursesFetch = await fetch('https://www.overpass-api.de/api/interpreter', {
       method: 'POST',
       headers: {
         'Accept': 'application/json',
         'Content-Type': 'application/json'
       },
-      body: "[out:json][timeout:25];(area[\"name\"=\"Toronto\"];)->.searchArea;nwr[\"leisure\"=\"golf_course\"](area.searchArea);out body;"
+      body: "[out:json][timeout:25];way(around:10000,43.6886058, -79.3004177)[\"leisure\"=\"golf_course\"];out body;"
     })
       .then(response => response.json())
       .then(data => { console.log(data); setCourses(data); setCourseLoading(false) })
       .catch(error => console.error(error));
     ;
-    console.log("The stuff", await courseFetch)
+    console.log("The stuff", await coursesFetch)
 
   }
-
   useEffect(() => { courseList() }, [courseLoading]);
 
   return (
@@ -48,7 +55,7 @@ export default function Login(props) {
           <CourseSelect
             setButtonTree={setButtonTree}
             navigation={navigation}
-            setPickedHole={setPickedHole}
+            setPickedCourse={setPickedCourse}
             courses={courses}
             setCourses={setCourses}
             courseLoading={courseLoading}
@@ -77,7 +84,7 @@ export default function Login(props) {
         {buttonTree[3] ? (
           <GameReview
             navigation={navigation}
-            pickedHole={pickedHole}
+            pickedCourse={pickedCourse}
             playerCount={playerCount}
             players={players}
           />
@@ -85,132 +92,11 @@ export default function Login(props) {
           <></>
         )}
       </View>
-    </View>
-  );
-}
-
-function CourseSelect(props) {
-  const { setButtonTree, buttonTree, setPickedHole, courses, setCourses, courseLoading } = props;
-  // console.log("Courses", courses.elements)
-  const elements = courses.elements;
-  console.log(courseLoading)
-  return (
-    <View>
-      <View>
-        <Text>Please choose a course:</Text>
-      </View>
-      {courseLoading ? <View><Text>Loading Text</Text></View> : elements.map((course, i) => {
-        {/* {courseStuff.map((course, i) => { */ }
-        // console.log(course)
-        return (
-          <View style={styles.course} key={i}>
-            <Button
-              key={i}
-              title={course.tags.name}
-              style={styles.button}
-              onPress={() => {
-                setPickedHole(course.id);
-                setButtonTree([false, true, false, false]);
-              }}
-            ></Button>
-            <View
-              style={{ flexDirection: "row", justifyContent: "space-evenly" }}>
-              <Text>{course.address}</Text>
-              <Text>Current weather</Text>
-            </View>
-          </View>
-        );
-      })}
-    </View>
-  );
-}
-
-
-function PlayerSelect(props) {
-  const { setButtonTree, buttonTree, setPlayerCount } = props;
-  return (
-    <View style={{ alignSelf: "center" }}>
-      <Text>How many players?</Text>
-      <View style={{ flexDirection: "row", justifyContent: "space-between" }}>
-        <Pressable
-          style={styles.playerButton}
-          onPress={() => {
-            setButtonTree([false, false, true, false]), setPlayerCount(1);
-          }}
-        >
-          <Text style={styles.playercount}>1</Text>
-        </Pressable>
-        <Pressable
-          style={styles.playerButton}
-          onPress={() => {
-            setButtonTree([false, false, true, false]), setPlayerCount(2);
-          }}
-        >
-          <Text style={styles.playercount}>2</Text>
-        </Pressable>
-        <Pressable
-          style={styles.playerButton}
-          onPress={() => {
-            setButtonTree([false, false, true, false]), setPlayerCount(3);
-          }}
-        >
-          <Text style={styles.playercount}>3</Text>
-        </Pressable>
-        <Pressable
-          style={styles.playerButton}
-          onPress={() => {
-            setButtonTree([false, false, true, false]), setPlayerCount(4);
-          }}
-        >
-          <Text style={styles.playercount}>4</Text>
-        </Pressable>
-      </View>
-    </View>
-  );
-}
-
-function PlayerNames(props) {
-  const {
-    setButtonTree,
-    buttonTree,
-    navigation,
-    playerCount,
-    players,
-    setPlayers,
-  } = props;
-
-  const playerNum = Array.from(
-    { length: playerCount },
-    (_, index) => index + 1
-  );
-
-  let playerNames = [];
-  return (
-    <View>
-      {playerNum.map((player, i) => {
-        if (i < playerCount) {
-          playerNames[i] = "Player " + (i + 1);
-          let playName = playerNames[i];
-          return (
-            <View style={styles.playerbox} key={i}>
-              <TextInput
-                style={styles.entername}
-                key={i}
-                defaultValue={playName}
-                clearTextOnFocus="true"
-                onChangeText={(playerNom) => {
-                  playerNames[i] = playerNom;
-                }}
-              ></TextInput>
-            </View>
-          );
-        }
-      })}
       <Button
         style={[styles.button, styles.goButton]}
-        title="Start Game!"
+        title="Reset"
         onPress={() => {
-          setButtonTree([false, false, false, true]), setPlayers(playerNames);
+          resetButtons();
         }}
       ></Button>
     </View>
@@ -218,62 +104,97 @@ function PlayerNames(props) {
 }
 
 function GameReview(props) {
-  const { pickedHole, playerCount, players, navigation } = props;
-  let course = courses.courses[pickedHole];
-  let holeCount = course.holes.length;
+  const { pickedCourse, playerCount, players, navigation, courses } = props;
+  const [course, setCourse] = useState([])
+  const [holesLoading, setHolesLoading] = useState(true)
+  async function getHoles() {
+    const courseFetch = await fetch('https://www.overpass-api.de/api/interpreter', {
+      method: 'POST',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json'
+      },
+      body: "[out:json][timeout:25];way(" + pickedCourse + ");map_to_area ->.golfcourse;way[\"golf\"=\"hole\"](area.golfcourse);(._;>;);out tags;"
+    })
+      .then(response => response.json())
+      .then(data => { console.log(data); setCourse(data); setHolesLoading(false) })
+      .catch(error => console.error(error));
+    ;
+    console.log("The hole stuff", await courseFetch)
+  }
+  // let course = pickedCourse;
+  useEffect(() => { getHoles() }, [holesLoading]);
 
-  let holeValue = Array.from({ length: holeCount }, (_, index) => 0);
-
-  const courseInfo = { name: course.name, address: course.address };
-  const playerInfo = players.map((player, i) => {
-    return {
-      player: player,
-      scores: holeValue,
-    };
-  });
-  const holeInfo = course.holes.map((hole) => {
-    return {
-      ...hole,
-      hole: hole.hole,
-      distance: hole.distance,
-      par: hole.par,
-    };
-  });
-  const scoreCard = {
-    course: courseInfo,
-    holes: holeInfo,
-    players: playerInfo,
-  };
-  return (
-    <View style={styles.recapBlock}>
-      <Text>Round details recap</Text>
-      <View style={styles.recapCourseBlock}>
-        <Text>{scoreCard.course.name}</Text>
-        <Text>{scoreCard.course.address}</Text>
-      </View>
-      <View style={styles.recapPlayersBlock}>
-        {scoreCard.players.map((golfer, i) => {
-          return <Text key={i}>{golfer.player}</Text>;
-        })}
-      </View>
-      <Pressable
-        style={[styles.button, styles.goButton]}
-        onPress={() =>
-          navigation.navigate("Scoreboard", {
-            // scoreCard: scoreCard,
-            course: courseInfo,
-            players: playerInfo,
-            holes: holeInfo,
-          })
+  let holes = [] 
+  if (!holesLoading) {
+      console.log("Let's go!", course)
+      course.elements.map((hole, i) => {
+        if (hole.type === "way") {
+          console.log(hole.tags)
         }
-      >
-        <Text>Start Game</Text>
-      </Pressable>
-    </View>
-  );
+      })
+    }
+
+  console.log(holes)
+  holesLoading ? console.log("Still going") : console.log("Done!")
+
+
+
+
+  // let holeCount = course.holes.length;
+
+  // let holeValue = Array.from({ length: holeCount }, (_, index) => 0);
+
+  // const courseInfo = { name: course.name, address: course.address };
+  // const playerInfo = players.map((player, i) => {
+  //   return {
+  //     player: player,
+  //     scores: holeValue,
+  //   };
+  // });
+  // const holeInfo = course.holes.map((hole) => {
+  //   return {
+  //     ...hole,
+  //     hole: hole.hole,
+  //     distance: hole.distance,
+  //     par: hole.par,
+  //   };
+  // });
+  // const scoreCard = {
+  //   course: courseInfo,
+  //   holes: holeInfo,
+  //   players: playerInfo,
+  // };
+  // return (
+  //   <View style={styles.recapBlock}>
+  //     <Text>Round details recap</Text>
+  //     <View style={styles.recapCourseBlock}>
+  //       <Text>{scoreCard.course.name}</Text>
+  //       <Text>{scoreCard.course.address}</Text>
+  //     </View>
+  //     <View style={styles.recapPlayersBlock}>
+  //       {scoreCard.players.map((golfer, i) => {
+  //         return <Text key={i}>{golfer.player}</Text>;
+  //       })}
+  //     </View>
+  //     <Pressable
+  //       style={[styles.button, styles.goButton]}
+  //       onPress={() =>
+  //         navigation.navigate("Scoreboard", {
+  //           // scoreCard: scoreCard,
+  //           course: courseInfo,
+  //           players: playerInfo,
+  //           holes: holeInfo,
+  //         })
+  //       }
+  //     >
+  //       <Text>Start Game</Text>
+  //     </Pressable>
+  //   </View>
+  // );
 }
 
-const styles = StyleSheet.create({
+export const styles = StyleSheet.create({
   button: {
     borderRadius: 20,
     elevation: 2,
