+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "@react-native-community/hooks": "^100.1.0",
+    "@react-native-masked-view/masked-view": "0.3.2",
     "@react-navigation/native": "^7.1.17",
     "@react-navigation/native-stack": "^7.3.26",
     "expo": "54.0.12",
@@ -18,6 +19,7 @@
     "react-dom": "19.2.0",
     "react-native": "0.81.4",
     "react-native-async-storage": "^0.0.1",
+    "react-native-gesture-handler": "~2.28.0",
     "react-native-safe-area-context": "5.6.1",
     "react-native-screens": "~4.16.0",
     "react-native-web": "~0.21.1"
